# Optimizer variadic elementwise ops