**Description**: Optimizer variadic elementwise ops

For cases like this, 12 layers, reusing the same weights. 

![image](https://user-images.githubusercontent.com/10530022/127147598-8c0dd2b8-ebf0-4f45-8b76-9553d6bfad63.png)

In PyTorch, after each PythonOpGrad executes, it will accumulate the grad into a buffer (except the first step), so 11 elementwise operator.

![image](https://user-images.githubusercontent.com/10530022/127149121-fb62f97a-d9e3-441f-87ea-85d0f59d7223.png)

In ORT, we will do the gradient sum with a Sum operator at the end, check the part the green arrow points to. 

![image](https://user-images.githubusercontent.com/10530022/127149342-9d605708-7338-49d3-80fb-4c089742d35d.png)

If we take a closer look: 

![image](https://user-images.githubusercontent.com/10530022/127149457-4520cbab-8913-4536-ab79-8fda83dc6afb.png)

This is a cuda memset + 12 binary element wise operator. 

Though we could do something to align our scheduling with PyTorch's for better memory usage. But this PR only trying to change "a cuda memset (1.12ms) + 12 binary element wise operator (47.278ms)" to "a cuda memcpy (2.75ms) + 11 binary element wise operator (43.16ms)".e.g. 48.398ms v.s.  45.91ms, 5.14% kernel overhead reduction for this case.



**Motivation and Context**
- Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here.