@pengwa 2021-09-27
close this, since PR https://github.com/microsoft/onnxruntime/pull/9186 will address the problem even better.